Changes state on submit
