+++ libs/application/templates/children-residence-change/src/lib/ChildrenResidenceChangeTemplate.ts
@@ -9,17 +9,13 @@ import {
 } from '@island.is/application/core'
 import { extractParentFromApplication } from './utils'
 import { assign } from 'xstate'
-import { error } from './messages/index'
-import * as z from 'zod'
+import { dataSchema } from './dataSchema'
 
-type Events =
-  | { type: DefaultEvents.ASSIGN }
-  | { type: DefaultEvents.SUBMIT }
+type Events = { type: DefaultEvents.ASSIGN } | { type: DefaultEvents.SUBMIT }
 
 enum States {
   DRAFT = 'draft',
   IN_REVIEW = 'inReview',
-  SUBMIT = 'submit'
 }
 
 enum Roles {
@@ -27,30 +23,7 @@ enum Roles {
   ParentB = 'parentB',
 }
 
-export const dataSchema = z.object({
-    approveExternalData: z.boolean().refine((v) => v, {
-      message: error.validation.approveChildrenResidenceChange.defaultMessage,
-    }),
-    selectChild: z
-      .array(z.string())
-      .min(1, { message: error.validation.selectChild.defaultMessage }),
-    email: z.string().email(error.validation.invalidEmail.defaultMessage),
-    phoneNumber: z
-      .string()
-      .min(7, { message: error.validation.invalidPhoneNumber.defaultMessage }),
-    confirmResidenceChangeInfo: z
-      .array(z.string())
-      .length(1, error.validation.approveChildrenResidenceChange.defaultMessage),
-    // selectDuration: z
-    //   .enum(['temporary', 'permanent'])
-    //   .optional()
-    //   .refine((v) => v, {
-    //     message: 'Velja þarf valmöguleika',
-    //   }),
-    approveTerms: z
-      .array(z.string())
-      .length(3, error.validation.approveTerms.defaultMessage),
-  })
+const applicationName = 'Umsókn um breytt lögheimili barns'
 
 const ChildrenResidenceChangeTemplate: ApplicationTemplate<
   ApplicationContext,
@@ -59,13 +32,13 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
 > = {
   type: ApplicationTypes.CHILDREN_RESIDENCE_CHANGE,
   name: 'Children residence change application',
-  dataSchema: dataSchema,
+  dataSchema,
   stateMachineConfig: {
     initial: States.DRAFT,
     states: {
       [States.DRAFT]: {
         meta: {
-          name: 'Umsókn um breytt lögheimili barns',
+          name: applicationName,
           progress: 0.33,
           roles: [
             {
@@ -75,14 +48,18 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
                   Promise.resolve(module.ChildrenResidenceChangeForm),
                 ),
               actions: [
-                { event: DefaultEvents.ASSIGN, name: 'Staðfesta', type: 'primary' },
+                {
+                  event: DefaultEvents.ASSIGN,
+                  name: 'Staðfesta',
+                  type: 'primary',
+                },
               ],
               write: 'all',
             },
           ],
         },
         on: {
-          SUBMIT: {
+          ASSIGN: {
             target: States.IN_REVIEW,
           },
         },
@@ -90,7 +67,7 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
       [States.IN_REVIEW]: {
         entry: 'assignToOtherParent',
         meta: {
-          name: 'Umsókn um breytt lögheimili barns',
+          name: applicationName,
           progress: 0.66,
           roles: [
             {
@@ -99,17 +76,9 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
                 import('../forms/ParentBForm').then((module) =>
                   Promise.resolve(module.ParentBForm),
                 ),
-              actions: [
-                { event: DefaultEvents.SUBMIT, name: 'Staðfesta', type: 'primary' },
-              ],
               write: 'all',
             },
           ],
-        },
-        on: {
-          SUBMIT: {
-            target: DefaultEvents.SUBMIT,
-          },
         },
       },
     },
@@ -119,13 +88,13 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
       assignToOtherParent: assign((context) => {
         const otherParent = extractParentFromApplication(context.application)
 
-          return {
-            ...context,
-            application: {
-              ...context.application,
-              assignees: [otherParent.ssn],
-            },
-          }
+        return {
+          ...context,
+          application: {
+            ...context.application,
+            assignees: [otherParent.ssn],
+          },
+        }
       }),
     },
   },
@@ -134,7 +103,6 @@ const ChildrenResidenceChangeTemplate: ApplicationTemplate<
     id: string,
     application: Application,
   ): ApplicationRole | undefined {
-    console.log({id})
     if (id === application.applicant) {
       return Roles.ParentA
     }
